requirements.txt
# Blue Team Assistant - Dependencies
# Author: Ugur Ates

# Core
requests>=2.28.0
pyyaml>=6.0
python-dateutil>=2.8.0

# File Analysis
pefile>=2023.2.7
yara-python>=4.3.0
python-magic>=0.4.27
ssdeep>=3.4
oletools>=0.60

# Email Analysis
dnspython>=2.3.0
mail-parser>=3.15.0

# Web Dashboard
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
jinja2>=3.1.0
pydantic>=2.5.0

# Cryptography (certificate validation)
cryptography>=41.0.0

# Optional: LLM Integration
# anthropic>=0.18.0
# openai>=1.0.0

# Optional: Advanced Analysis
# capa>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aioresponses>=0.7.4
